requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==40.0.1
Django==4.1
django-cors-headers==3.14.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
pycparser==2.21
pytz==2023.2
sqlparse==0.4.3
tzdata==2023.2
